add verificacao produto
